event driven widgets; bootstrap thrown in for fun
Switch library widgets to be event driven:

Instead of widgets that each watch for state changes on popstate and re-run
unless we are able to short-circuit, each widget runs an init function the
first time it gets rendered and listens for custom events to update when needed.
A single watcher maps url state changes to custom events. This allows better
decoupling of most of the UI elements from eachother, as well as cutting down
on unnecessary processing on each popstate. It should also let us get rid of
several ugly workarounds on mobile currently necessary because of interaction
between different browser history usage. Finally, it is better suited to
underlying data being more heavily cached and updating using the sync API.

Also switch to bootstrap to see how much jqueryUI specific code is intertwined
where it shouldn't be.