Workaround for a stale attachment URL #451 #452
Timestamp is now taken after the request has been processed, not before.
This way if a request is fired 60 seconds later, it is guaranteed to
receive a fresh URL. The downside is that before that second request
comes back, the URL rendered is stale. To fix this a server-side change
is needed: zotero/dataserver#115