Cache meta and templates for 24h. Resolve #166
Template requests and meta requests, both general and item-type
specific, are now cached for 24 hours to improve performance.
Time of the last request is recorded in localStorage, then, if cache is
considered fresh, we use ask browser to use `force-cache` caching
strategy which will go to the disk cache. If that fails (or cache is
not fresh), network request is performed.