Rework excludeCollections as disabledCollections
Previously collections tree accepted includeCollections and
excludeCollections params which would filter collections that appeared
in the tree, same as libraries tree does for libraries. However it's
confusing for a collection node to disappear from the tree in picker
mode so instead we switch to mechanism where all collections still
appear but some cannot be navigated to.