Fix a rare regression where the table may display ghost rows
When using a URL parameter to navigate directly to an item in the trash,
if collections are also trashed and there are more items than one page
of results, the web library would display ghost rows.

This regression is related to changes in d2821f3f and 2ae0a6ff.