Fix picker navigation logic to respect `pickerSkip` condition. #625
Before 677376e, users on touch devices could use a dialog to move an attachment a library without file editing capabilities, leading to a server error (though different from the one described in #625). After 677376e, while that issue was resolved, it became possible to navigate into such a library and select a collection within it. This commit resolves that problem.